Clean Up
Fixed Game Breaking bug

added header comments
removed unused functions

+++ LoserGameForLosers/AICommands.cpp
@@ -83,14 +83,14 @@ void AICommands::RandRangeCombat::execute()
 					block_time_->stop();
 				}
 			}
-			else if (rand() % 1000 < block_rand_range_)
+			else if (static_cast<Uint32>(rand() % 1000) < block_rand_range_)
 			{
 				block_->init();
 				block_time_->start();
 				block_time_total_ = rand() % 1000 + 500;
 			}
 
-			if (rand() % 1000 < push_rand_range_)
+			if (static_cast<Uint32>(rand() % 1000) < push_rand_range_)
 			{
 				push_->init();
 				bot_->get_component<ScriptComponent>().empty_add(push_->get_trunk());
@@ -98,7 +98,7 @@ void AICommands::RandRangeCombat::execute()
 			
 			if (bot_->get_component<TransformComponent>().get_horizontal_distance(player_) < KICK_LENGTH)
 			{
-				if (rand() % 1000 < kick_rand_range_)
+				if (static_cast<Uint32>(rand() % 1000) < kick_rand_range_)
 				{
 					kick_->init();
 					bot_->get_component<ScriptComponent>().empty_add(kick_->get_trunk());
