+++ LoserGameForLosers/TextureAtlasComponent.h
@@ -3,9 +3,9 @@
 #include "TextureSlot.h"
 
 /**
- * TextureAtlasComponent 
- * -------------------------
- * Creates and Updates images from dynamically sized vector of "Slots" created from data 
+ * @author David Wadsworth
+ *
+ * Creates and Updates images from dynamically sized vector of "Slots" (either animated or static) created from data 
  */
 
 class TextureAtlasComponent : public Component
@@ -25,9 +25,9 @@ public:
 	// Standard image to store in slot 
 	int create_image_slot(int sprite_id, int x, int y, int rotation = 0, SDL_RendererFlip flip = SDL_FLIP_NONE, int slot_id = -1);
 	void update_image_slot(int slot_id, int updated_tex, int rotation = 0, SDL_RendererFlip flip = SDL_FLIP_NONE);
-	void update_position_and_scaling(int slot_id, Vector2D position, float scale);
+	void update_position_and_scaling(int slot_id, Vector2D position, double scale);
 	void update_position(int slot_id, Vector2D position);
-	void update_scaling(int slot_id, float scale);
+	void update_scaling(int slot_id, double scale);
 	void update_dimensions(int slot_id, int width, int height);
 	void update_height(int slot_id, int height);
 	void update_width(int slot_id, int width);
